microservices/service_gateway/requirements.txt
Flask
Flask-session
SQLAlchemy
Flask-SQLAlchemy
mysqlclient
Flask-RESTful
python-decouple